Registro de usuarios #8 Permitir login de usuarios contra la base de datos #3

+++ src/main/java/tk/srubio/adoptix/model/ProvinceRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProvinceRepository extends JpaRepository<Province, Integer> {
+public interface ProvinceRepository extends JpaRepository<Province, Byte> {
 
 	public List<Province> findAllByOrderByNameAsc();
